Changed file creation mechanism
Was able to generate a crash even just touching the files. Thus, file
size so far doesn't seem to matter much for the purposes of unison
crashing. Adjusted to just touch files to try to boost the speed through
the test some.

Also, did some testing with different values of file_loops and
dupe_loops. Got crashes but later in the process when going down to
4096 files, 4 loops and 8192 files, 3 loops.